Added user login persistence, not working maybe because of slow fetch speed
